+++ packages/pytest-simcore/src/pytest_simcore/simcore_services.py
@@ -4,12 +4,11 @@
 
 import asyncio
 import logging
-from typing import Any, Dict, List
+from typing import Dict, List
 
 import aiohttp
 import pytest
 import tenacity
-from distributed import Client
 from yarl import URL
 
 from .helpers.utils_docker import get_service_published_port
@@ -58,33 +57,6 @@ async def simcore_services(services_endpoint: Dict[str, URL], monkeypatch) -> No
         env_prefix = service.upper().replace("-", "_")
         monkeypatch.setenv(f"{env_prefix}_HOST", endpoint.host)
         monkeypatch.setenv(f"{env_prefix}_PORT", str(endpoint.port))
-
-
-@pytest.fixture(scope="function")
-async def dask_scheduler_service(simcore_services, monkeypatch) -> Dict[str, Any]:
-    # the dask scheduler has a UI for the dashboard and a secondary port for the API
-    # simcore_services fixture already ensure the dask-scheduler is up and running
-    dask_scheduler_api_port = get_service_published_port(
-        "dask-scheduler", target_ports=[8786]
-    )
-    # override the port
-    monkeypatch.setenv("DASK_SCHEDULER_PORT", f"{dask_scheduler_api_port}")
-    return {"host": "127.0.0.1", "port": dask_scheduler_api_port}
-
-
-@pytest.fixture(scope="function")
-def dask_client(dask_scheduler_service: Dict[str, Any]) -> Client:
-
-    client = Client(
-        f"{dask_scheduler_service['host']}:{dask_scheduler_service['port']}"
-    )
-    yield client
-    client.close()
-
-
-@pytest.fixture(scope="function")
-def dask_sidecar_service(dask_client: Client) -> None:
-    dask_client.wait_for_workers(1, 30)
 
 
 # HELPERS --
